requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
cloc==0.2.5
colored==2.2.4
idna==3.10
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
faiss_cpu==1.7.4
google_api_python_client==2.142.0
google_auth_oauthlib==1.0.0
plotly
protobuf==5.29.1
pytest==7.4.0
python-dotenv==1.0.1
PyYAML==6.0.2
streamlit==1.38.0
streamlit_authenticator==0.3.3